Refactor WebSocket Handler
