+++ PoGo.NecroBot.CLI/WebSocketHandler/ActionCommands/TransferPokemonHandler.cs
@@ -4,10 +4,10 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using SuperSocket.WebSocket;
-using PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands.Tasks;
+using PoGo.NecroBot.CLI.WebSocketHandler.GetCommands.Tasks;
 using PoGo.NecroBot.Logic.State;
 
-namespace PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands
+namespace PoGo.NecroBot.CLI.WebSocketHandler.ActionCommands
 {
     public class TransferPokemonHandler : IWebSocketRequestHandler
     {
@@ -20,7 +20,7 @@ namespace PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands
 
         public async Task Handle(ISession session, WebSocketSession webSocketSession, dynamic message)
         {
-            await TransferPokemonTask.Execute(session, webSocketSession, (ulong)message.PokemonId, (string)message.RequestID);
+            await Logic.Tasks.TransferPokemonTask.Execute(session, (ulong)message.PokemonId);
         }
     }
 }
