+++ PoGo.NecroBot.CLI/WebSocketHandler/EncodingHelper.cs
@@ -10,7 +10,6 @@ namespace PoGo.NecroBot.CLI.WebSocketHandler
 {
     class EncodingHelper
     {
-
         public static string Serialize(dynamic evt)
         {
             var jsonSerializerSettings = new JsonSerializerSettings { TypeNameHandling = TypeNameHandling.All };
