+++ PoGo.NecroBot.CLI/WebSocketHandler/GetCommands/Events/EggListResponce.cs
@@ -4,18 +4,19 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands.Events
+namespace PoGo.NecroBot.CLI.WebSocketHandler.GetCommands.Events
 {
-    class ItemListResponce : IWebSocketResponce
+    public class EggListResponce : IWebSocketResponce
     {
-        public ItemListResponce(dynamic data,string requestID)
+        public EggListResponce(dynamic data, string requestID)
         {
-            Command = "ItemListWeb";
+            Command = "EggListWeb";
             Data = data;
             RequestID = requestID;
         }
         public string RequestID { get; private set; }
         public string Command { get; private set; }
         public dynamic Data { get; private set; }
+        
     }
 }
