+++ PoGo.NecroBot.CLI/WebSocketHandler/GetCommands/Events/TrainerProfileResponce.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands.Events
+namespace PoGo.NecroBot.CLI.WebSocketHandler.GetCommands.Events
 {
     class TrainerProfileResponce : IWebSocketResponce
     {
