+++ PoGo.NecroBot.CLI/WebSocketHandler/GetCommands/GetEggListHandler.cs
@@ -1,4 +1,4 @@
-﻿using PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands.Tasks;
+﻿using PoGo.NecroBot.CLI.WebSocketHandler.GetCommands.Tasks;
 using PoGo.NecroBot.Logic.State;
 using SuperSocket.WebSocket;
 using System;
@@ -7,21 +7,20 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands
+namespace PoGo.NecroBot.CLI.WebSocketHandler.GetCommands
 {
-    class GetTrainerProfileHandler : IWebSocketRequestHandler
+    class GetEggListHandler : IWebSocketRequestHandler
     {
-
         public string Command { get; private set; }
 
-        public GetTrainerProfileHandler()
+        public GetEggListHandler()
         {
-            Command = "GetTrainerProfile";
+            Command = "GetEggList";
         }
 
         public async Task Handle(ISession session, WebSocketSession webSocketSession, dynamic message)
         {
-            await GetTrainerProfileTask.Execute(session, webSocketSession, (string)message.RequestID);
+            await GetEggListTask.Execute(session, webSocketSession, (string)message.RequestID);
         }
     }
 }
