+++ PoGo.NecroBot.CLI/WebSocketHandler/GetCommands/GetItemsListHandler.cs
@@ -1,4 +1,4 @@
-﻿using PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands.Tasks;
+﻿using PoGo.NecroBot.CLI.WebSocketHandler.GetCommands.Tasks;
 using PoGo.NecroBot.Logic.State;
 using SuperSocket.WebSocket;
 using System;
@@ -7,20 +7,21 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands
+namespace PoGo.NecroBot.CLI.WebSocketHandler.GetCommands
 {
-    class GetEggListHandler : IWebSocketRequestHandler
+    class GetItemsListHandler : IWebSocketRequestHandler
     {
         public string Command { get; private set; }
 
-        public GetEggListHandler()
+        public GetItemsListHandler()
         {
-            Command = "GetEggList";
+            Command = "GetItemsList";
         }
 
         public async Task Handle(ISession session, WebSocketSession webSocketSession, dynamic message)
         {
-            await GetEggListTask.Execute(session, webSocketSession, (string)message.RequestID);
+            await GetItemListTask.Execute(session, webSocketSession, (string)message.RequestID);
         }
+
     }
 }
