+++ PoGo.NecroBot.CLI/WebSocketHandler/GetCommands/GetPokemonListHandler.cs
@@ -4,23 +4,23 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using SuperSocket.WebSocket;
-using PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands.Tasks;
+using PoGo.NecroBot.CLI.WebSocketHandler.GetCommands.Tasks;
 using PoGo.NecroBot.Logic.State;
 
-namespace PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands
+namespace PoGo.NecroBot.CLI.WebSocketHandler.GetCommands
 {
-    public class GetPokemonSettingsHandler : IWebSocketRequestHandler
+    public class GetPokemonListHandler : IWebSocketRequestHandler
     {
         public string Command { get; private set;}
 
-        public GetPokemonSettingsHandler()
+        public GetPokemonListHandler()
         {
-            Command = "GetPokemonSettings";
+            Command = "GetPokemonList";
         }
 
         public async Task Handle(ISession session, WebSocketSession webSocketSession, dynamic message)
         {
-            await GetPokemonSettingsTask.Execute(session, webSocketSession, (string)message.RequestID);
+            await GetPokemonListTask.Execute(session, webSocketSession, (string)message.RequestID);
         }
     }
 }
