+++ PoGo.NecroBot.CLI/WebSocketHandler/GetCommands/GetTrainerProfileHandler.cs
@@ -1,4 +1,4 @@
-﻿using PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands.Tasks;
+﻿using PoGo.NecroBot.CLI.WebSocketHandler.GetCommands.Tasks;
 using PoGo.NecroBot.Logic.State;
 using SuperSocket.WebSocket;
 using System;
@@ -7,21 +7,21 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands
+namespace PoGo.NecroBot.CLI.WebSocketHandler.GetCommands
 {
-    class GetItemsListHandler : IWebSocketRequestHandler
+    class GetTrainerProfileHandler : IWebSocketRequestHandler
     {
+
         public string Command { get; private set; }
 
-        public GetItemsListHandler()
+        public GetTrainerProfileHandler()
         {
-            Command = "GetItemsList";
+            Command = "GetTrainerProfile";
         }
 
         public async Task Handle(ISession session, WebSocketSession webSocketSession, dynamic message)
         {
-            await GetItemListTask.Execute(session, webSocketSession, (string)message.RequestID);
+            await GetTrainerProfileTask.Execute(session, webSocketSession, (string)message.RequestID);
         }
-
     }
 }
