+++ PoGo.NecroBot.CLI/WebSocketHandler/GetCommands/Helpers/EggListWeb.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands.Helpers
+namespace PoGo.NecroBot.CLI.WebSocketHandler.GetCommands.Helpers
 {
     class EggListWeb
     {
