+++ PoGo.NecroBot.CLI/WebSocketHandler/GetCommands/Helpers/TrainerProfileWeb.cs
@@ -6,7 +6,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace PoGo.NecroBot.CLI.WebSocketHandler.BasicGetCommands.Helpers
+namespace PoGo.NecroBot.CLI.WebSocketHandler.GetCommands.Helpers
 {
     class TrainerProfileWeb
     {
